requirements.txt
python-telegram-bot 
pytube 
pydub
yt-dlp